# docs/swagger.yaml
definitions:
...
  DatabaseServicev1.Card:
    properties:
      createdAt:
        description: '* Дата создания сущности в базе данных'
        type: string
      cvv:
        description: '* CVV код карты'
        type: integer
      date:
        description: '* Дата до которой активна карта'
        type: string
      fullName:
        description: '* ФИО с банковской карты'
        type: string
      id:
        description: '* ID банковской карты в базе данных'
        type: integer
      number:
        description: '* Номер карты'
        type: string
      updatedAt:
        description: '* Дата последнего обновления сущности в базе данных'
        type: string
      userId:
        description: '* ID пользователя которому принадлежит данная карта'
        type: integer
    type: object
  DatabaseServicev1.CardCompany:
    properties:
      companyId:
        description: '* ID компании которой принадлежит данная карта'
        type: integer
      createdAt:
        description: '* Дата создания сущности в базе данных'
        type: string
      cvv:
        description: '* CVV код карты'
        type: integer
      date:
        description: '* Дата до которой активна карта'
        type: string
      fullName:
        description: '* ФИО с банковской карты'
        type: string
      id:
        description: '* ID банковской карты в базе данных'
        type: integer
      number:
        description: '* Номер карты'
        type: string
      updatedAt:
        description: '* Дата последнего обновления сущности в базе данных'
        type: string
    type: object
  DatabaseServicev1.CardsCompaniesResponse:
    properties:
      cards:
        description: '* Массив банковских карт компаний'
        items:
          $ref: '#/definitions/DatabaseServicev1.CardCompany'
        type: array
    type: object
...
  DatabaseServicev1.CreateCardCompanyRequest:
    properties:
      companyId:
        description: '* ID компании которой принадлежит данная карта'
        type: integer
      cvv:
        description: '* CVV код банковской карты'
        type: integer
      date:
        description: '* Дата до которой активна карта'
        type: string
      fullName:
        description: '* ФИО с банковской карты'
        type: string
      number:
        description: '* Номер банковской карты'
        type: string
    type: object
  DatabaseServicev1.CreateCardRequest:
    properties:
      cvv:
        description: '* CVV код карты'
        type: integer
      date:
        description: '* Дата до которой активна карта'
        type: string
      fullName:
        description: '* ФИО с банковской карты'
        type: string
      number:
        description: '* Номер карты'
        type: string
      userId:
        description: '* ID пользователя кому принадлежит карта'
        type: integer
    type: object
  DatabaseServicev1.CreateCompanyRequest:
    properties:
      address:
        description: '* Адрес офиса компании'
        type: string
      card:
        allOf:
        - $ref: '#/definitions/DatabaseServicev1.CardCompany'
        description: '* Банковская карта компании'
      inn:
        description: '* ИНН юридического лица'
        type: string
      kpp:
        description: '* КПП юридического лица'
        type: string
      okpo:
        description: '* ОКПО предприятия/организации'
        type: string
      phone:
        description: '* Номер телефона компании'
        type: string
      site:
        description: '* Сайт компании'
        type: string
      title:
        description: '* Название компании'
        type: string
      userId:
        description: '* ID пользователя к которому относится данная компания'
        type: integer
    type: object
...
  DatabaseServicev1.HTTPCodes:
    properties:
      code:
        description: '* HTTP код ответа'
        type: integer
    type: object
...
  server.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  server.LoginRequest:
    properties:
      password:
        type: string
      phone:
        type: string
    type: object
  server.LoginResponse:
    properties:
      token:
        type: string
    type: object
...
  server.RegistrationRequest:
    properties:
      card:
        $ref: '#/definitions/DatabaseServicev1.CreateCardRequest'
      company:
        $ref: '#/definitions/DatabaseServicev1.CreateCompanyRequest'
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      type:
        type: integer
      username:
        type: string
    type: object
info:
  contact: {}
  description: Сервер маршрутизации
  title: API Gateway
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Авторизация пользователя
      parameters:
      - description: Данные для авторизации пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/server.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Авторизация
      tags:
      - Authentication
  /api/v1/auth/registration:
    post:
      consumes:
      - application/json
      description: Регистрация нового пользователя пользователя
      parameters:
      - description: Данные для регистрации
        in: body
        name: user
        schema:
          $ref: '#/definitions/server.RegistrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Регистрация пользователя
      tags:
      - Authentication
  /api/v1/card/company:
    get:
      consumes:
      - application/json
      description: Банковская карта компании в базе данных
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.CardsCompaniesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Банковская карта компании
      tags:
      - CardCompany
    post:
      consumes:
      - application/json
      description: Создание банковской карты компании
      parameters:
      - description: Сущность банковской карты компании
        in: body
        name: card
        schema:
          $ref: '#/definitions/DatabaseServicev1.CreateCardCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.Card'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Создание банковской карты компании
      tags:
      - CardCompany
    put:
      consumes:
      - application/json
      description: Обновление банковской карты компании
      parameters:
      - description: Модель для обновления
        in: body
        name: card
        schema:
          $ref: '#/definitions/DatabaseServicev1.CardCompany'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.CardCompany'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Обновление банковской карты компании
      tags:
      - CardCompany
  /api/v1/card/company/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление банковской карты компании по ID
      parameters:
      - description: ID банковской карты компании
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.HTTPCodes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Удаление банковской карты компании
      tags:
      - CardCompany
    get:
      consumes:
      - application/json
      description: Поиск банковской карты компании по ID
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.CardCompany'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Поиск банковской карты компании
      tags:
      - CardCompany
  /api/v1/card/company/deleteModel:
    post:
      consumes:
      - application/json
      description: Удаляет банковскую карту компании опираясь на всю сущность модели
      parameters:
      - description: Сущность банковской карты компании
        in: body
        name: card
        schema:
          $ref: '#/definitions/DatabaseServicev1.CardCompany'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.HTTPCodes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Удаление банковской карты компании по модели
      tags:
      - CardCompany
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"